Allow partial pruning in mempool
Since micro blocks are not validated via consensus, different
nodes may have different state of mempool. In this case it is possible
to create a situation where the same UTXO is used twice for two different
transactions on two different nodes. Of course, the only one of such
transactions will be processed and included to the block, depending on
the leader election order.

This patch removes removes from mempool all transactions which become
orphan after received a new micro block. Before this patch the usage of
the same UTXO in two different transactions caused panic of the node

Closes #755