Remove `peer` and `sequence_number` from FloodsubMessage
Both `peer` and `sequence_numbers` fields effectively changed the hash
of FloodsubMessage even if the payload was the same. It led to multiple
transmission of the same message for some cases in pBFT and Wallet.
We don't want to pay that price to solve the problem which doesn't
exist for our case. Zap it.

This patch also removes Topic, TopicBuilder, TopicHash, and entire topic.rs
because it was completely useless piece of code which did nothing.
Zap it.

This patch breaks compatibility with FloodSub implementation from
js-libp2p and go-libp2p, but who cares?